Add files via upload
GiveHediffSeverity is like the vanilla comp, except it adds to existing severity instead of replacing it.
CorpseDamagerToHD is very versatile; it stacks a given hediff's severity according to the type of corpses around the thing with the comp. You can stack different hediffs according to whether there are mechs/human/bug/animal corpses around, damage them or not, only find one corpse/all corpses per check, and so on. Auto eat corpses, repair by melding with other dead mechs, get weaker around dead corpses.
InitialHediffIfFaction is mostly there for Biotech mechs, so that only enemy/player mechs get a starting hediff.
SummonFromHediffSeverity spawns a thing and consumes the hediff's severity to spawn it.
Verb_CooldownJump is MVCF's jump but checks for an internal cooldown; it's only there because I didn't realize that an ability must also be integrated to a ThinkTree to be used. Maybe potentially useful?
ConeShoot is very cool; it's basically the FireSpew comp from Biotech except usable for other things.